[Refactor] Commit1: Refactor `maybe_compile`, make it only apply to single function
The previous `maybe_compile` functionality was overly complex, as it was responsible for both determining whether individual functions should be compiled, in-place replacement, and runtime checking. This had the following drawbacks:

1. maybe_compile wrapped functions with a layer, and the wrapper contained some dynamic conditional branches at runtime, which could potentially affect the compile behavior.

2. `torch.compile` itself is quite fragile, and with the wrapper in place, it's difficult to correctly handle all compilable objects, such as normal functions, methods, etc.

3. `maybe_compile` maintained a global dictionary to determine which functions need to be compiled and which don't. However, determining the key for functions is a rather complex matter. For example, the qualname of a normal function can serve as the key, while methods require supplementing with an id, making it difficult to add & remove compile objects.

Therefore, this commit restricts the functionality of `maybe_compile`, making it only responsible for serving as a wrapper for module-level functions. It has only one function: to ensure that when other modules import the function, what they actually import is the same `MaybeCompile` object. This facilitates global in-place replacement at runtime for all functions that may need to be compiled.


ghstack-source-id: f2f70ef59e06d2347601c93ac69ecb3560a92d46
Pull-Request: https://github.com/InternLM/xtuner/pull/1317